Move "Visitor" entity
It was (mis)placed in the use case, now it's a official domain entitity.
